refactor: 물건 관련 DTO 및 서비스 파일명 수정
