+++ app/src/main/java/com/example/swapit/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.activity.compose.setContent
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.rememberNavController
 import com.example.swapit.domain.repository.LoginRepository
-import com.example.swapit.domain.repository.ShoppingRepository
+import com.example.swapit.domain.repository.ProductRepository
 import com.example.swapit.ui.auth.LoginManager
 import com.example.swapit.ui.auth.LoginViewModel
 import com.example.swapit.ui.navigation.NavigationModule
@@ -27,18 +27,9 @@ class MainActivity : ComponentActivity() {
                             LoginManager(this),
                         ),
                 )
-            val shoppingViewModel: ShoppingViewModel =
-                viewModel(
-                    factory =
-                        ShoppingViewModel.factory(
-                            ShoppingRepository.instance(),
-                        ),
-                )
-
             navigationModule.NavigationGraph(
                 navController,
                 loginViewModel,
-                shoppingViewModel,
             )
         }
     }
