--- app/src/main/java/com/example/swapit/data/datasource/RemoteMyProductSelectDataSource.kt
@@ -1,11 +1,0 @@
-package com.example.swapit.data.datasource
-
-import com.example.swapit.data.datasource.remote.dto.response.BaseResponse
-import com.example.swapit.data.datasource.remote.dto.response.product.ProductResponse
-import com.example.swapit.data.datasource.remote.service.ProductService
-
-class RemoteMyProductSelectDataSource(private val myProductSelectService: ProductService) {
-    suspend fun myProductSelectResponse(): BaseResponse<List<ProductResponse>> {
-        return myProductSelectService.myProductSelectResponse()
-    }
-}
