+++ app/src/main/java/com/example/swapit/data/datasource/RemoteSwapDataSource.kt
@@ -4,9 +4,9 @@ import com.example.swapit.data.datasource.remote.dto.request.swap.SwapRequest
 import com.example.swapit.data.datasource.remote.dto.response.BaseResponse
 import com.example.swapit.data.datasource.remote.service.SwapService
 
-class RemoteSwapRequestDataSource(private val swapRequestService: SwapService) {
+class RemoteSwapDataSource(private val swapService: SwapService) {
     suspend fun swapRequest(swapRequest: SwapRequest): BaseResponse<Long> =
-        swapRequestService.swapRequest(
+        swapService.swapRequest(
             swapRequest,
         )
 }
