--- app/src/main/java/com/example/swapit/data/repository/DefaultMyProductSelectRepository.kt
@@ -1,19 +1,0 @@
-package com.example.swapit.data.repository
-
-import com.example.swapit.data.datasource.RemoteMyProductSelectDataSource
-import com.example.swapit.data.datasource.remote.dto.response.BaseResponse
-import com.example.swapit.data.datasource.remote.dto.response.product.ProductResponse
-import com.example.swapit.data.mapper.toDomain
-import com.example.swapit.domain.model.shopping.ShoppingProduct
-import com.example.swapit.domain.repository.MyProductSelectRepository
-
-class DefaultMyProductSelectRepository(private val remoteSource: RemoteMyProductSelectDataSource) :
-    MyProductSelectRepository {
-    override suspend fun myProductSelectResults(): List<ShoppingProduct> {
-        return remoteSource.myProductSelectResponse().results.map { it.toDomain() }
-    }
-
-    override suspend fun myProductSelectResponse(): BaseResponse<List<ProductResponse>> {
-        return remoteSource.myProductSelectResponse()
-    }
-}
