+++ app/src/main/java/com/example/swapit/ui/shopping/SearchBarButton.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
@@ -22,7 +23,7 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.example.swapit.R
-import com.example.swapit.domain.repository.ShoppingRepository
+import com.example.swapit.domain.repository.ProductRepository
 import com.example.swapit.ui.navigation.NavItem
 import com.example.swapit.ui.theme.Gray3
 import com.example.swapit.ui.theme.Gray6
@@ -78,6 +79,7 @@ fun SearchBarButton(
 @Preview(showBackground = true)
 fun SearchBarButtonPreview() {
     SwapitTheme {
-        SearchBarButton(Modifier, rememberNavController(), viewModel = ShoppingViewModel(ShoppingRepository.instance()))
+        SearchBarButton(Modifier, rememberNavController(), viewModel = ShoppingViewModel(
+            ProductRepository.instance(LocalContext.current)))
     }
 }
