+++ app/src/main/java/com/example/swapit/ui/shopping/ShoppingScreen.kt
@@ -15,11 +15,12 @@ import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
-import com.example.swapit.domain.repository.ShoppingRepository
+import com.example.swapit.domain.repository.ProductRepository
 import com.example.swapit.ui.component.AppBar
 import com.example.swapit.ui.component.BottomNavigationBar
 import com.example.swapit.ui.navigation.NavItem
@@ -91,6 +92,6 @@ fun ShoppingScreen(
 @Composable
 fun ProductScreenPreview() {
     SwapitTheme {
-        ShoppingScreen(rememberNavController(), ShoppingViewModel(ShoppingRepository.instance()))
+        ShoppingScreen(rememberNavController(), ShoppingViewModel(ProductRepository.instance(LocalContext.current)))
     }
 }
