+++ app/src/main/java/com/example/swapit/ui/shopping/detail/select/MyProductSelectScreen.kt
@@ -10,11 +10,12 @@ import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarColors
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
-import com.example.swapit.domain.repository.MyProductSelectRepository
-import com.example.swapit.domain.repository.SwapRequestRepository
+import com.example.swapit.domain.repository.ProductRepository
+import com.example.swapit.domain.repository.SwapRepository
 import com.example.swapit.ui.component.BackButton
 import com.example.swapit.ui.theme.BackgroundColor
 import com.example.swapit.ui.theme.Paddings
@@ -38,7 +39,7 @@ fun MyProductSelectScreen(
             TextSection(navController)
             ProductListSection(
                 myProductSelectionViewModel = viewModel,
-                swapViewModel = SwapProductViewModel(repository = SwapRequestRepository.instance()),
+                swapViewModel = SwapProductViewModel(repository = SwapRepository.instance()),
                 targetProductId = targetProductId,
             )
         }
@@ -72,7 +73,7 @@ fun MyProductAppBar(navController: NavHostController) {
 fun SwapRequestScreenPreview() {
     MyProductSelectScreen(
         rememberNavController(),
-        viewModel = MyProductSelectViewModel(repository = MyProductSelectRepository.instance()),
+        viewModel = MyProductSelectViewModel(repository = ProductRepository.instance(LocalContext.current)),
         targetProductId = 0,
     )
 }
